Remove IBinarySerializerFactory interface and reorganize SAGE.IO code
