+++ SAGESharpTests/SAGESharpTests.csproj
@@ -77,13 +77,13 @@
     <Compile Include="EqualityUtilsTests.cs" />
     <Compile Include="IO\BinaryReaderWrapperTests.cs" />
     <Compile Include="IO\BinarySerializableSerializerTests.cs" />
+    <Compile Include="IO\BinarySerializerTests.cs" />
     <Compile Include="IO\BinaryWriterWrapperTests.cs" />
     <Compile Include="IO\EdgeTests.cs" />
     <Compile Include="IO\ListNodeTests.cs" />
     <Compile Include="IO\OffsetNodeTests.cs" />
     <Compile Include="IO\PrimitiveTypeDataNodeTests.cs" />
     <Compile Include="IO\StringDataNodeTests.cs" />
-    <Compile Include="IO\TreeBasedBinarySerializerFactoryTests.cs" />
     <Compile Include="IO\TreeBinarySerializerTests.cs" />
     <Compile Include="IO\TreeBuilderTests.cs" />
     <Compile Include="IO\TreeReaderTests.cs" />
